yolov5/FastAPI/requirements.txt
Package                         Version
------------------------------- --------------------
absl-py                         1.4.0
actionlib                       1.12.1
angles                          1.9.12
anyio                           3.6.2
base_local_planner              1.16.7
bondpy                          1.8.5
cachetools                      5.3.0
camera_calibration              1.15.2
camera_calibration_parsers      1.11.13
catkin                          0.7.29
certifi                         2022.12.7
charset-normalizer              3.1.0
click                           8.1.3
cmake                           3.26.0
contourpy                       1.0.7
controller_manager              0.18.4
controller_manager_msgs         0.18.4
cv_bridge                       1.13.1
cycler                          0.11.0
diagnostic_analysis             1.9.7
diagnostic_common_diagnostics   1.9.7
diagnostic_updater              1.9.7
drone_wrapper                   1.3.10
dynamic_reconfigure             1.6.5
fastapi                         0.95.0
filelock                        3.10.2
fonttools                       4.39.2
gazebo_plugins                  2.8.7
gazebo_ros                      2.8.7
gencpp                          0.6.6
geneus                          2.2.6
genlisp                         0.4.16
genmsg                          0.5.17
gennodejs                       2.0.1
genpy                           0.6.16
gitdb                           4.0.10
GitPython                       3.1.31
google-auth                     2.16.2
google-auth-oauthlib            0.4.6
grpcio                          1.51.3
h11                             0.14.0
idna                            3.4
image_geometry                  1.13.1
importlib-metadata              6.1.0
importlib-resources             5.12.0
interactive-markers             1.11.5
Jinja2                          3.1.2
joint_state_publisher           1.12.15
joint_state_publisher_gui       1.12.15
kdl-parser-py                   1.13.3
kiwisolver                      1.4.4
laser_geometry                  1.6.7
lit                             16.0.0
Markdown                        3.4.2
MarkupSafe                      2.1.2
matplotlib                      3.7.1
mavros                          1.14.0
message_filters                 1.14.13
moveit-core                     1.0.11
moveit_ros_planning_interface   1.0.11
mpmath                          1.3.0
networkx                        3.0
numpy                           1.24.2
nvidia-cublas-cu11              11.10.3.66
nvidia-cuda-cupti-cu11          11.7.101
nvidia-cuda-nvrtc-cu11          11.7.99
nvidia-cuda-runtime-cu11        11.7.99
nvidia-cudnn-cu11               8.5.0.96
nvidia-cufft-cu11               10.9.0.58
nvidia-curand-cu11              10.2.10.91
nvidia-cusolver-cu11            11.4.0.1
nvidia-cusparse-cu11            11.7.4.91
nvidia-nccl-cu11                2.14.3
nvidia-nvtx-cu11                11.7.91
oauthlib                        3.2.2
opencv-python                   4.7.0.72
packaging                       23.0
pandas                          1.5.3
Pillow                          9.4.0
pip                             23.0.1
protobuf                        4.22.1
psutil                          5.9.4
py-trees                        0.6.9
pyasn1                          0.4.8
pyasn1-modules                  0.2.8
pydantic                        1.10.7
pyparsing                       3.0.9
python-dateutil                 2.8.2
python_qt_binding               0.4.4
pytz                            2022.7.1
PyYAML                          6.0
qt-dotgraph                     0.4.2
qt-gui                          0.4.2
qt-gui-cpp                      0.4.2
qt-gui-py-common                0.4.2
requests                        2.28.2
requests-oauthlib               1.3.1
resource_retriever              1.12.7
rosbag                          1.14.13
rosboost-cfg                    1.14.9
rosclean                        1.14.9
roscreate                       1.14.9
rosgraph                        1.14.13
roslaunch                       1.14.13
roslib                          1.14.9
roslint                         0.11.2
roslz4                          1.14.13
rosmake                         1.14.9
rosmaster                       1.14.13
rosmsg                          1.14.13
rosnode                         1.14.13
rosparam                        1.14.13
rospy                           1.14.13
rosserial_python                0.8.0
rosservice                      1.14.13
rostest                         1.14.13
rostopic                        1.14.13
rosunit                         1.14.9
roswtf                          1.14.13
rqt_action                      0.4.9
rqt_bag                         0.5.1
rqt_bag_plugins                 0.5.1
rqt_console                     0.4.9
rqt_controller_manager          0.18.4
rqt_dep                         0.4.9
rqt_drone_teleop                1.3.10
rqt_ez_publisher                0.5.0
rqt_graph                       0.4.11
rqt_ground_robot_teleop         1.3.10
rqt_gui                         0.5.3
rqt_gui_py                      0.5.3
rqt_image_view                  0.4.16
rqt_joint_trajectory_controller 0.17.2
rqt_joint_trajectory_plot       0.0.5
rqt_launch                      0.4.8
rqt_launchtree                  0.2.0
rqt_logger_level                0.4.8
rqt-moveit                      0.5.10
rqt_msg                         0.4.8
rqt_multiplot                   0.0.10
rqt_nav_view                    0.5.7
rqt_paramedit                   1.0.1
rqt_play_motion_builder         1.0.2
rqt_plot                        0.4.13
rqt_pose_view                   0.5.8
rqt_publisher                   0.4.8
rqt_py_common                   0.5.3
rqt_py_console                  0.4.8
rqt_py_trees                    0.3.1
rqt-reconfigure                 0.5.4
rqt_robot_dashboard             0.5.7
rqt-robot-monitor               0.5.14
rqt_robot_steering              0.5.10
rqt_rotors                      2.2.3
rqt_runtime_monitor             0.5.7
rqt-rviz                        0.7.0
rqt_service_caller              0.4.8
rqt_shell                       0.4.9
rqt_srv                         0.4.8
rqt_tf_tree                     0.6.0
rqt_top                         0.4.8
rqt_topic                       0.4.11
rqt_web                         0.4.8
rsa                             4.9
rviz                            1.13.29
scipy                           1.10.1
seaborn                         0.12.2
sensor-msgs                     1.12.8
setuptools                      67.4.0
six                             1.16.0
smach                           2.0.1
smach_ros                       2.0.1
smclib                          1.8.5
smmap                           5.0.0
sniffio                         1.3.0
srdfdom                         0.5.2
starlette                       0.26.1
sympy                           1.11.1
tensorboard                     2.12.0
tensorboard-data-server         0.7.0
tensorboard-plugin-wit          1.8.1
tf                              1.12.1
tf_conversions                  1.12.1
tf2_geometry_msgs               0.6.5
tf2_kdl                         0.6.5
tf2_py                          0.6.5
tf2_ros                         0.6.5
thop                            0.1.1.post2209072238
topic_tools                     1.14.13
torch                           2.0.0
torchvision                     0.15.1
tqdm                            4.65.0
triton                          2.0.0
typing_extensions               4.5.0
unique_id                       1.0.6
urdfdom-py                      0.4.6
urllib3                         1.26.15
uvicorn                         0.21.1
Werkzeug                        2.2.3
wheel                           0.38.4
xacro                           1.13.18
zipp                            3.15.0
